backend/requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Authentication
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
django-allauth==0.60.1
google-auth==2.23.4

# Database
mysqlclient==2.2.1
pymysql==1.1.0

# File storage
Pillow==10.0.1
django-storages==1.14.2
boto3==1.28.62

# Payments
stripe==7.11.0

# Environment variables
python-dotenv==1.0.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3


# QR Codes
qrcode==7.4.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Documentation
drf-yasg==1.21.7

# Security
python-jose==3.3.0
cryptography==41.0.7





dj-rest-auth[social]==5.0.2
django-allauth[socialaccount]==0.60.1
python-social-auth==0.3.6
social-auth-app-django==5.4.2
